refactoring based on no coins going into OwnedNFTView
